finished all tests - 2 external bugs remaining
