+++ pennylane/templates/layers/particle_conserving_u2.py
@@ -134,6 +134,17 @@ class ParticleConservingU2(Operation):
             layers = 1
             params = qml.init.particle_conserving_u2_normal(layers, qubits)
             print(cost_fn(params))
+
+        **Parameter shape**
+
+        The shape of the weights argument can be computed by the static method
+        :meth:`~.ParticleConservingU2.shape` and used when creating randomly
+        initialised weight tensors:
+
+        .. code-block:: python
+
+            shape = ParticleConservingU2.shape(n_layers=2, n_wires=2)
+            weights = np.random.random(size=shape)
     """
 
     num_params = 1
